Step 6 -- Steps 5 and 6 need try-catch
Note an idea in the Class Admin and Appointment to have Objects intercommunicating instead of Strings.

Test the code inside delete functions (Admin) and see if it truly deletes the passed objects (you should have a NullReferenceException).

Plus, Check the Appointments Class in order to add some function (even a fixed one, like the Patients and Doctor objects) and continue the tests.

Shall we connect a Database within the program?